requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
google-generativeai
python-dotenv
psycopg2-binary
gunicorn